+++ server/src/database/interest.test.ts
@@ -1,0 +1,6 @@
+import { expect, test } from "bun:test";
+import * as core from "./interest";
+
+test("list", async () => {
+  expect(await core.all()).toSatisfy(() => true);
+});
