+++ server/src/seeds/data/subjects.ts
@@ -1,4 +1,4 @@
-const subjects = [
+export const subjects = [
   {
     group: "Computer Science",
     subjects: ["機械学習", "CPU アーキテクチャ", "型システム", "分散処理"],
@@ -8,5 +8,3 @@ const subjects = [
     subjects: ["Lean4"],
   },
 ];
-
-export default subjects satisfies Array<{ group: string; subjects: string[] }>;
